LazyImports.jl

LazyImports.jl

LazyImports.@lazy_import

LazyImports.@shim_import

LazyImports.@lazy_import — Macro.

@lazy_import Module
@lazy_import Module = UUID

Construct a module-like object for Module. Accessing Module.f imports Module and then get f from Module. The UUID can be omitted if Module is listed in [deps] or [extras] of Project.toml file of the package using @lazy_import.

source

LazyImports.@shim_import — Macro.

@shim_import Module
@shim_import Module = UUID

This acts like @lazy_import except that Module.f works only if Module is already loaded.

source